CURD

+++ app/interface/internal/service/problem.go
@@ -33,6 +33,7 @@ func (s *ProblemService) ListProblems(ctx context.Context, req *v1.ListProblemsR
 			Name:          v.Name,
 			SubmitCount:   int32(v.SubmitCount),
 			AcceptedCount: int32(v.AcceptedCount),
+			Status:        int32(v.Status),
 			CreatedAt:     timestamppb.New(v.CreatedAt),
 			UpdatedAt:     timestamppb.New(v.UpdatedAt),
 		})
@@ -42,6 +43,9 @@ func (s *ProblemService) ListProblems(ctx context.Context, req *v1.ListProblemsR
 
 // 题目详情
 func (s *ProblemService) GetProblem(ctx context.Context, req *v1.GetProblemRequest) (*v1.Problem, error) {
+	// if ok := s.uc.HasPermission(ctx, int(req.Id), "view"); !ok {
+	// 	return nil, v1.ErrorPermissionDenied("permission denied")
+	// }
 	data, err := s.uc.GetProblem(ctx, int(req.Id))
 	if err != nil {
 		return nil, err
@@ -49,6 +53,7 @@ func (s *ProblemService) GetProblem(ctx context.Context, req *v1.GetProblemReque
 	resp := &v1.Problem{
 		Id:            int32(data.ID),
 		Name:          data.Name,
+		Status:        int32(data.Status),
 		MemoryLimit:   int32(data.MemoryLimit),
 		TimeLimit:     int32(data.TimeLimit),
 		SubmitCount:   int32(data.SubmitCount),
@@ -91,16 +96,23 @@ func (s *ProblemService) CreateProblem(ctx context.Context, req *v1.CreateProble
 
 // 创建题目
 func (s *ProblemService) UpdateProblem(ctx context.Context, req *v1.UpdateProblemRequest) (*v1.Problem, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	_, err := s.uc.UpdateProblem(ctx, &biz.Problem{
 		ID:          int(req.Id),
 		TimeLimit:   req.TimeLimit,
 		MemoryLimit: req.MemoryLimit,
+		Status:      int(req.Status),
 	})
 	return nil, err
 }
 
 // 获取题目描述列表
 func (s *ProblemService) ListProblemStatements(ctx context.Context, req *v1.ListProblemStatementsRequest) (*v1.ListProblemStatementsResponse, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	res, count := s.uc.ListProblemStatements(ctx, req)
 	resp := new(v1.ListProblemStatementsResponse)
 	resp.Total = count
@@ -125,6 +137,9 @@ func (s *ProblemService) GetProblemStatement(ctx context.Context, req *v1.GetPro
 
 // 创建题目描述
 func (s *ProblemService) CreateProblemStatement(ctx context.Context, req *v1.CreateProblemStatementRequest) (*v1.ProblemStatement, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	res, err := s.uc.CreateProblemStatement(ctx, &biz.ProblemStatement{
 		ProblemID: int(req.Id),
 		Language:  req.Language,
@@ -139,6 +154,9 @@ func (s *ProblemService) CreateProblemStatement(ctx context.Context, req *v1.Cre
 
 // 更新题目描述
 func (s *ProblemService) UpdateProblemStatement(ctx context.Context, req *v1.UpdateProblemStatementRequest) (*v1.ProblemStatement, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	res, err := s.uc.UpdateProblemStatement(ctx, &biz.ProblemStatement{
 		ID:        int(req.Sid),
 		ProblemID: int(req.Id),
@@ -158,6 +176,9 @@ func (s *ProblemService) UpdateProblemStatement(ctx context.Context, req *v1.Upd
 
 // DeleteProblemStatement 删除题目描述
 func (s *ProblemService) DeleteProblemStatement(ctx context.Context, req *v1.DeleteProblemStatementRequest) (*v1.ProblemStatement, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	err := s.uc.DeleteProblemStatement(ctx, int(req.Sid))
 	if err != nil {
 		return nil, err
@@ -167,6 +188,9 @@ func (s *ProblemService) DeleteProblemStatement(ctx context.Context, req *v1.Del
 
 // ListProblemTests 获取题目测试点列表
 func (s *ProblemService) ListProblemTests(ctx context.Context, req *v1.ListProblemTestsRequest) (*v1.ListProblemTestsResponse, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	data, count := s.uc.ListProblemTests(ctx, req)
 	resp := new(v1.ListProblemTestsResponse)
 	resp.Count = count
@@ -191,6 +215,9 @@ func (s *ProblemService) GetProblemTest(ctx context.Context, req *v1.GetProblemT
 
 // 创建题目测试点
 func (s *ProblemService) CreateProblemTest(ctx context.Context, req *v1.CreateProblemTestRequest) (*v1.ProblemTest, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	res, err := s.uc.CreateProblemTest(ctx, &biz.ProblemTest{
 		ProblemID:        int(req.Id),
 		InputFileContent: req.InputFileContent,
@@ -232,6 +259,9 @@ func (s *ProblemService) UploadProblemTest(ctx http.Context) error {
 
 // 更新题目测试点
 func (s *ProblemService) UpdateProblemTest(ctx context.Context, req *v1.UpdateProblemTestRequest) (*v1.ProblemTest, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	s.uc.UpdateProblemTest(ctx, &biz.ProblemTest{
 		ID:        req.Tid,
 		ProblemID: int(req.Id),
@@ -243,12 +273,18 @@ func (s *ProblemService) UpdateProblemTest(ctx context.Context, req *v1.UpdatePr
 
 // 删除题目测试点
 func (s *ProblemService) DeleteProblemTest(ctx context.Context, req *v1.DeleteProblemTestRequest) (*emptypb.Empty, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	err := s.uc.DeleteProblemTest(ctx, int64(req.Id), req.Tid)
 	return &emptypb.Empty{}, err
 }
 
 // 获取题目文件列表
 func (s *ProblemService) ListProblemFiles(ctx context.Context, req *v1.ListProblemFilesRequest) (*v1.ListProblemFilesResponse, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	res, count := s.uc.ListProblemFiles(ctx, req)
 	resp := new(v1.ListProblemFilesResponse)
 	resp.Total = count
@@ -266,6 +302,9 @@ func (s *ProblemService) ListProblemFiles(ctx context.Context, req *v1.ListProbl
 
 // 获取题目文件详情
 func (s *ProblemService) GetProblemFile(ctx context.Context, req *v1.GetProblemFileRequest) (*v1.ProblemFile, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	res, err := s.uc.GetProblemFile(ctx, int(req.Sid))
 	if err != nil {
 		return nil, err
@@ -283,6 +322,9 @@ func (s *ProblemService) GetProblemFile(ctx context.Context, req *v1.GetProblemF
 
 // 创建题目文件
 func (s *ProblemService) CreateProblemFile(ctx context.Context, req *v1.CreateProblemFileRequest) (*v1.ProblemFile, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	res, err := s.uc.CreateProblemFile(ctx, &biz.ProblemFile{
 		ProblemID: int(req.Id),
 		Content:   req.Content,
@@ -300,6 +342,9 @@ func (s *ProblemService) CreateProblemFile(ctx context.Context, req *v1.CreatePr
 
 // 更新题目文件
 func (s *ProblemService) UpdateProblemFile(ctx context.Context, req *v1.UpdateProblemFileRequest) (*v1.ProblemFile, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	s.uc.UpdateProblemFile(ctx, &biz.ProblemFile{
 		ID:      int(req.Sid),
 		Name:    req.Name,
@@ -311,6 +356,9 @@ func (s *ProblemService) UpdateProblemFile(ctx context.Context, req *v1.UpdatePr
 
 // 删除题目文件
 func (s *ProblemService) DeleteProblemFile(ctx context.Context, req *v1.DeleteProblemFileRequest) (*v1.ProblemFile, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	s.uc.DeleteProblemFile(ctx, int(req.Sid))
 	return nil, nil
 }
@@ -322,6 +370,9 @@ func (s *ProblemService) RunProblemFile(ctx context.Context, req *v1.RunProblemF
 
 // 获取题目文件列表
 func (s *ProblemService) ListProblemStdCheckers(ctx context.Context, req *v1.ListProblemStdCheckersRequest) (*v1.ListProblemStdCheckersResponse, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	res, _ := s.uc.ListProblemFiles(ctx, &v1.ListProblemFilesRequest{
 		FileType: "checker",
 	})
@@ -339,16 +390,25 @@ func (s *ProblemService) ListProblemStdCheckers(ctx context.Context, req *v1.Lis
 }
 
 func (s *ProblemService) VerifyProblem(ctx context.Context, req *v1.VerifyProblemRequest) (*emptypb.Empty, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	err := s.uc.VerifyProblem(ctx, int(req.Id))
 	return &emptypb.Empty{}, err
 }
 
 func (s *ProblemService) UpdateProblemChecker(ctx context.Context, req *v1.UpdateProblemCheckerRequest) (*emptypb.Empty, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	err := s.uc.UpdateProblemChecker(ctx, int(req.Id), int(req.CheckerId))
 	return &emptypb.Empty{}, err
 }
 
 func (s *ProblemService) GetProblemVerification(ctx context.Context, req *v1.GetProblemVerificationRequest) (*v1.ProblemVerification, error) {
+	if ok := s.uc.HasPermission(ctx, int(req.Id), "update"); !ok {
+		return nil, v1.ErrorPermissionDenied("permission denied")
+	}
 	resp := new(v1.ProblemVerification)
 	res, err := s.uc.GetProblemVerification(ctx, int(req.Id))
 	if err == nil {
